Remove Laravel references
Strips a few Laravel references, changes Liobot to Octobot, and removes the laravel-specific !relsched and !wiki commands.

The !rules command links to a gist that also contains a lot of references we won't want, but it's useful too. I'll probably fork it and update before I change the link here.